+++ tests/testapp/settings.py
@@ -7,6 +7,7 @@ import os
 DEBUG = True
 
 DATABASES = {"default": {"ENGINE": "django.db.backends.sqlite3", "NAME": "imagefield"}}
+DEFAULT_AUTO_FIELD = "django.db.models.AutoField"
 
 INSTALLED_APPS = [
     "django.contrib.auth",