+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Change log
 ~~~~~~~~~~~~~~~
 
 - Renamed the main branch to ``main``.
+- Reformatted the frontend code using prettier and checked it using ESLint.
 
 
 `0.13`_ (2021-02-03)